requirements.txt
torch>=1.9.0
transformers>=4.20.0
pandas>=1.3.0
numpy==1.21.0
datasets>=2.0.0
evaluate>=0.3.0
accelerate>=0.20.0
peft>=0.4.0
biopython>=1.79
scikit-learn>=1.0.0
bitsandbytes>=0.39.0
galore-torch>=0.1.0
google-cloud-bigquery>=3.35.0
typer[all]>=0.12.0
